+++ build.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-sudo DOCKER_BUILDKIT=1 docker build --rm -t ankit/devbox:1.0 . 2>&1 | tee logs/build.log
+sudo DOCKER_BUILDKIT=1 docker build --rm -t ankit/devbox:1.1 . 2>&1 | tee logs/build.log